server/pyproject.toml
[tool.poetry]
name = "inrutu-museum"
version = "0.1.0"
description = "Backend для виртуального музея вычислительной техники"
authors = ["Ivan <ivan@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.0"
djangorestframework = "^3.14.0"
django-filter = "^23.0"
drf-spectacular = "^0.26.0"
psycopg2-binary = "^2.9.0"
python-decouple = "^3.8"
pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-django = "^4.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88
